updated start file and some other stuff :^)

+++ start.sh
@@ -14,6 +14,54 @@ git=true
 py3=true
 py=true
 
+function edit_config {
+    editor="none"
+    type vim >/dev/null 2>&1 && editor="vim"
+    type nano >/dev/null 2>&1 && editor="nano"
+    if [ ! $editor = "none" ]; then
+        $editor config.json
+    fi
+}
+
+function dl_config {
+    wget https://raw.githubusercontent.com/zekroTJA/serverManager/master/config_ex.json -o /dev/null
+    mv config_ex.json config.json
+}
+
+if [ ! -z $1 ]; then
+    if [ $1 = "help" ] || [ $1 = "--help" ]; then
+        echo "Usage: bash start.sh <command>"
+        echo ""
+        echo "  Option           Description"
+        echo " --------------------------------------------------------------"
+        echo "  help             Dispaly this help message"
+        echo "  conf | config    Open editor with config"
+        echo "  noupdate         Start without checking for update"
+        echo "  disable-update   Permanently disable update check on start"
+        echo "  enable-update    Re-enable auto-update at start"
+        echo "  reset            Reset setting for updates and package check"
+        echo "  nostart          Only update without starting src/main.py"
+        echo "                   Actually for development purpose"
+        echo ""
+        exit
+    fi
+
+    if [ $1 = "conf" ] || [ $1 = "config" ]; then
+        if [ ! -f config.json ]; then
+            dl_config
+        fi
+        edit_config
+        exit
+    fi
+
+    if [ $1 = "reset" ]; then
+        rm DISABLEUPDATE >/dev/null 2>&1
+        rm CHECKEDPACKAGES >/dev/null 2>&1
+        echo "Reset DISABLEUPDATE and CHECKEDPACKAGES"
+        exit
+    fi
+fi
+
 # Testing if 'git' command works
 type git >/dev/null 2>&1 || git=false
 
@@ -34,39 +82,46 @@ fi
 # Checks if the required python packages are
 # installed
 type $check_cmd >/dev/null 2>&1 && {
-    echo "Testing for required packages..."
-    printf "  - psutil: "
-    echo $(pip3 list --format=columns) | grep "psutil" --silent
-    if [ $? = 1 ]
+    if [ ! -f CHECKEDPACKAGES ]
     then
-        printf "will be installed... "
-        $check_cmd install psutil >/dev/null
-        echo "installed"
-    else
-        echo "installed"
+        echo "Testing for required packages..."
+        printf "  - psutil: "
+        echo $(pip3 list --format=columns) | grep "psutil" --silent
+        if [ ! -z $1 ]; then
+            if [ $? = 1 ]
+            then
+                printf "will be installed... "
+                $check_cmd install psutil >/dev/null
+                echo "installed"
+            else
+                echo "installed"
+            fi
+        fi
+        echo ""
+        echo " " >> CHECKEDPACKAGES
     fi
-    echo ""
 }
-
-
-# When 'disable-updat' parameter is given, there will be
-# created a file named 'DISABLEUPDATE' in root directory
-# of this script to signal that updated should not pulled
-# automatically on startup
-if [ "$1" = "disable-update" ]
-then
-    echo " " >> DISABLEUPDATE
-fi
-
-# If parameter 'enable-update' is given and the file
-# 'DISABLEUPDATE' is existing, the file will be deleted
-# so that the tool will automatically pull updates from
-# origin repository
-if [ "$1" = "enable-update" ]
-then
-    if [ -f DISABLEUPDATE ]
+    
+if [ ! -z $1 ]; then
+    # When 'disable-updat' parameter is given, there will be
+    # created a file named 'DISABLEUPDATE' in root directory
+    # of this script to signal that updated should not pulled
+    # automatically on startup
+    if [ "$1" = "disable-update" ]
     then
-        rm DISABLEUPDATE
+        echo " " >> DISABLEUPDATE
+    fi
+
+    # If parameter 'enable-update' is given and the file
+    # 'DISABLEUPDATE' is existing, the file will be deleted
+    # so that the tool will automatically pull updates from
+    # origin repository
+    if [ "$1" = "enable-update" ]
+    then
+        if [ -f DISABLEUPDATE ]
+        then
+            rm DISABLEUPDATE
+        fi
     fi
 fi
 
@@ -91,20 +146,20 @@ else
         echo "# You can re-enable it with parameter 'enable-update' #" #
         echo "#######################################################" #
         echo ""                                                        #
-        echo "[Starting in 3 seconds...]"                              #
-        sleep 3                                                        #
+        echo "[Starting in 2 seconds...]"                              #
+        sleep 2                                                        #
         ################################################################
     else
         if [ ! "$1" = "noupdate" ]
         then
             if [ ! -d .git ]
             then
-                echo "------------------------ ATTENTION ------------------------------"
+                echo "-------------------- ATTENTION --------------------------"
                 echo "This is not a cloned repository."
                 echo "This folder will be deleted and cloned from repository."
-                read -p "Continue? (y/n)" res
+                read -p "Continue? (y/n) " res
                 if [ "$res" = "n" ]; then exit; fi
-                echo "-----------------------------------------------------------------"
+                echo "---------------------------------------------------------"
                 echo "Cloning repository..."
                 currdir=$PWD
                 cd ..
@@ -112,18 +167,34 @@ else
                 git clone https://github.com/zekroTJA/serverManager.git $currdir
                 cd $currdir
                 mv config_ex.json config.json
+                edit_config
             else
                 if [ ! -f config.json ] && [ -f config_ex.json ]; then
                     mv config_ex.json config.json
+                    edit_config
                 fi
                 echo "Pulling update from origin repository..."
                 echo "-----------------------------------------"
                 git pull origin master
                 echo "-----------------------------------------"
-            echo "Completed updating repository."
+                echo "Completed updating repository."
             fi
         fi
     fi
+fi
+
+if [ ! -f config.json ]
+then
+    echo "------------------------------- ATTENTION ------------------------------------"
+    echo "No 'config.json' found!"
+    echo "Do you want to download the default config from:"
+    echo "https://raw.githubusercontent.com/zekroTJA/serverManager/master/config_ex.json"
+    read -p "(y/n) " res
+    if [ "$res" = "n" ]; then
+        exit 1
+    fi
+    dl_config
+    edit_config
 fi
 
 # Now, if python3 is installed, the tool will start the main script with
@@ -131,8 +202,10 @@ fi
 # with this version, which will - depending on the python version -
 # fail with a very high probability
 
-if [ $1 = "nostart" ]; then
-    exit 1
+if [ ! -z $1 ]; then
+    if [ $1 = "nostart" ]; then
+        exit 1
+    fi
 fi
 
 if $py3
